backend/requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
weasyprint
twilio
openai
gunicorn
